Accelerate property checking via bbox
With bbox command, we can black box modules interface (that is, ignore
logic design inside a module, and let jg to stimulate interface's port
directly).
After that, we can assume mem is ideal while not over constraint because
now we do not need to consider how is bus works anymore.